cnss2: Add X86 support
With this change, platform driver can load and unload successfully
on X86 platform. CONFIG_CNSS2_X86 is introduced for this purpose.

In this change, it changes to 32bit DMA only. Firmware dump is
captured via dev_coredump way.

In this change, ASPM is also disabled.

With this change, the rddm is captured in the cnss2 way but in
ath12k format.

The benefit is the sync mechanism between cnss2 and cld driver
remains same.

Signed-off-by: Carl Huang <quic_cjhuang@quicinc.com>

Change-Id: I3c53a2caf9d032644cf7305ae6778b41122b18fb