cnss_utils: Wait for initialization of daemon config
There is scenario where cnss-daemon is got connected with
cnss-plat-ipc module but message to initialize daemon
config is not yet received from cnss-daemon. At the same
time cnss2 module checks for DMS support in daemon config
which is not yet initialized. As use-nv-mac is set from DTSI,
DMS support must be enabled. Since it is not found enabled,
cnss2 asserts.
To fix this, along with cnss-daemon connection check also
wait for message from cnss-daemon which initializes daemon
config.

Change-Id: I2579e7e9040c0ea7286c265488e2050652c5cc21
CRs-Fixed: 3517692