cnss2: fix cnss2 driver reload fail with WLAN attached to PCIe switch
If WLAN is attached to PCIe switch, it's possible that the DSP <-> EP
link is disabled when reloading the cnss2 driver. This case can lead
to a failure in obtaining the PCI device ID, preventing the cnss2 driver
from being reloaded. To fix this issue, ensure that the DSP <-> EP link
is enabled when getting the PCI device ID fail

Change-Id: Icc51a96d2f61c92743a27a8e9be1ccb4e9b772e1
CRs-Fixed: 3848272