cnss2: set mhi buffer length to 8k
When allocate memory for mhi buffer, it needs order-4 memory.
There won't be so many large memory chunk and page allocation
failure will happen on LE projects. Set the mhi buffer length
to 8k so that only order-1 memory is required. Reducing this
size will not impact the message communication with FW.

Change-Id: I0b45184d9f96c20e89439be642185f3bc4502b1f
CRs-Fixed: 3983268